:hammer: Refactored utils for dataloading, validation, and tiled
